+++ doc/releasenotes.rst
@@ -14,6 +14,7 @@ Version 1.10
   added to the 64-bit view of the registry.
 * Fixed an error when using wheels which install files into the same package,
   such as ``PyQt5`` and ``PyQtChart``.
+* Issue a warning when we can't find the cache directory on Windows.
 
 Version 1.9
 -----------